@@ -12,28 +12,29 @@
     ],
     "autoload": {
         "psr-4": {
-            "Overblog\\DataLoaderBundle\\": ""
-        },
-        "exclude-from-classmap": [
-            "/Tests/"
-        ]
+            "Overblog\\DataLoaderBundle\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Overblog\\DataLoaderBundle\\Tests\\": "tests/"
+        }
     },
     "config" : {
         "bin-dir": "bin",
         "sort-packages": true
     },
     "require": {
-        "php": ">=7.2",
-        "overblog/dataloader-php": "^0.5.0 || ^0.6.0",
-        "symfony/dependency-injection": "^4.3 || ^5.0"
+        "php": "^8.0",
+        "overblog/dataloader-php": "^0.7.0",
+        "symfony/dependency-injection": "^5.3 || ^6.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.1 || ^5.5 || ^6.5 || ^7.5 || ^8.5",
+        "phpunit/phpunit": "^9.5.10",
         "react/promise": "^2.5",
-        "symfony/dependency-injection": "^4.3 || ^5.0",
-        "symfony/framework-bundle": "^4.3 || ^5.0",
-        "symfony/phpunit-bridge": "^4.3 || ^5.0",
-        "symfony/yaml": "^4.3 || ^5.0"
+        "sensio/framework-extra-bundle": "^6.2",
+        "symfony/phpunit-bridge": "^6.0",
+        "symfony/yaml": "^5.3 || ^6.0"
     },
     "extra": {
         "branch-alias": {
